Reorganize e2e tests
We have e2e tests organized a bit illogically, this attempts to fix it:

- test-e2e/complete/ - complete workflow tests which tests across
  multiple esy commands.
- test-e2e/common/ - some common tests.
- test-e2e/esy-build/ - tests for `esy build` invocation
- test-e2e/esy-install/ - tests for `esy install` invocation
- test-e2e/esy-COMMAND-NAME.test.js - tests for `esy COMMAND-NAME`
  invocation.
