+++ test-e2e/common/build-errors.test.js
