+++ test-e2e/common/solve-errors.test.js
