+++ test-e2e/complete/symlink-workflow.test.js
