+++ test-e2e/esy-build-CMD.test.js
@@ -3,7 +3,7 @@
 const os = require('os');
 const path = require('path');
 
-const helpers = require('../test/helpers');
+const helpers = require('./test/helpers');
 const {packageJson, dir, file, dummyExecutable, buildCommand} = helpers;
 
 helpers.skipSuiteOnWindows();
