+++ test-e2e/esy-build/augment-path.test.js
