+++ test-e2e/esy-build/same-dep-transitive-and-direct.test.js
