+++ test-e2e/esy-install/basic-npm.test.js
