+++ test-e2e/esy-install/pkg-integrity.test.js
