+++ test-e2e/esy-install/resolutions.test.js
