+++ test-e2e/esy-install/sources.test.js
